Edit flow inprogress
