+++ apps/front-end/src/pages/front-end/ag-edit/choose-subjects/subjectDetails.js
