+++ apps/front-end/src/pages/front-end/ag-edit/choose-subjects/subjectForm.js
@@ -1,7 +1,7 @@
-import React from "react";
+import React, { useEffect } from "react";
 import Form from "@rjsf/core";
 import validator from "@rjsf/validator-ajv8";
-import schema1 from "../parts/schema.js";
+import schema1 from "./schema.js";
 import {
   Alert,
   Box,
@@ -13,9 +13,11 @@ import {
   Radio,
   Stack,
   VStack,
+  Checkbox,
+  Pressable,
 } from "native-base";
 import CustomRadio from "../../../../component/CustomRadio.js";
-import Steper from "../../../../component/Steper";
+import Steper from "../../../../component/Steper.js";
 import {
   facilitatorRegistryService,
   geolocationRegistryService,
@@ -33,8 +35,10 @@ import {
   BodyMedium,
   changeLanguage,
   enumRegistryService,
-  benificiaryRegistoryService,
+  updateSchemaEnum,
 } from "@shiksha/common-lib";
+
+//updateSchemaEnum
 import moment from "moment";
 import { useNavigate } from "react-router-dom";
 import Clipboard from "component/Clipboard.js";
@@ -44,7 +48,7 @@ import {
   FieldTemplate,
   ObjectFieldTemplate,
   ArrayFieldTitleTemplate,
-} from "../../../../component/BaseInput";
+} from "../../../../component/BaseInput.js";
 import { useScreenshot } from "use-screenshot-hook";
 
 const CustomR = ({ options, value, onChange, required }) => {
@@ -81,7 +85,7 @@ const RadioBtn = ({ options, value, onChange, required }) => {
         w="75%"
         maxW="300px"
       >
-        {items.map((item) => (
+        {items?.map((item) => (
           <Radio key={item?.value} value={item?.value} size="lg">
             {item?.label}
           </Radio>
@@ -91,13 +95,69 @@ const RadioBtn = ({ options, value, onChange, required }) => {
   );
 };
 
+const CheckBoxes = ({ options, value, onChange, required }) => {
+  const [mainValue, setMainValue] = React.useState([]);
+  const items = options?.enumOptions;
+  const handleCheckboxChange = (isChecked, values) => {
+    const count = mainValue.find((e) => e != values);
+    console.log(count);
+    if (!count) {
+      const newData = [...mainValue, values];
+      console.log(...mainValue, values, isChecked, "sa");
+
+      setMainValue(newData);
+      onChange(newData);
+    } else {
+      const newData = mainValue.filter((e) => e != values);
+      setMainValue(newData);
+      onChange(newData);
+    }
+  };
+  return (
+    <Box
+      colorScheme="black"
+      defaultValue={value}
+      accessibilityLabel="choose multiple items"
+    >
+      <Stack
+        direction={{
+          base: "column",
+          md: "row",
+        }}
+        alignItems={{
+          base: "flex-start",
+          md: "center",
+        }}
+        space={4}
+        w="75%"
+        maxW="300px"
+      >
+        {items?.map((item) => (
+          <Pressable
+            key={item?.value}
+            value={item?.value}
+            bg={mainValue.includes(item?.value) ? "red.100" : "white"}
+            onPress={
+              (e) => {
+                handleCheckboxChange(e, item?.value);
+              }
+
+              //    onChange(e || []);
+            }
+          >
+            {item?.label}
+          </Pressable>
+        ))}
+      </Stack>
+    </Box>
+  );
+};
+
 // App
-export default function App({ ip, id }) {
+export default function App({ facilitator, ip, onClick }) {
   const [page, setPage] = React.useState();
   const [pages, setPages] = React.useState();
-  const [cameraData, setCameraData] = React.useState([]);
   const [schema, setSchema] = React.useState({});
-  const [cameraSelection, setCameraSelection] = React.useState(0);
   const [cameraModal, setCameraModal] = React.useState(false);
   const [credentials, setCredentials] = React.useState();
   const [cameraUrl, setCameraUrl] = React.useState();
@@ -105,16 +165,16 @@ export default function App({ ip, id }) {
   const [addBtn, setAddBtn] = React.useState(t("YES"));
   const formRef = React.useRef();
   const uplodInputRef = React.useRef();
-  const [formData, setFormData] = React.useState({});
+  const [formData, setFormData] = React.useState(facilitator);
   const [errors, setErrors] = React.useState({});
   const [alert, setAlert] = React.useState();
   const [yearsRange, setYearsRange] = React.useState([1980, 2030]);
   const [lang, setLang] = React.useState(localStorage.getItem("lang"));
   const navigate = useNavigate();
-  /*  const { form_step_number } = facilitator;
-  if (form_step_number && parseInt(form_step_number) >= 7) {
+  const { form_step_number } = facilitator;
+  if (form_step_number && parseInt(form_step_number) >= 13) {
     navigate("/dashboard");
-  } */
+  }
 
   window.onbeforeunload = function () {
     return false;
@@ -138,25 +198,13 @@ export default function App({ ip, id }) {
     getImage();
   }, [page, credentials]);
 
-  //getting data
-  React.useEffect(async () => {
-    const qData = await benificiaryRegistoryService.getOne(id);
-    console.log(qData.result, "asd");
-    setFormData(qData.result);
-  }, []);
-
-  React.useEffect(() => {
-    if (page == "5") {
-      console.log(cameraData);
-      if (cameraData.length <= 3) {
-        setCameraModal(true);
-      } else {
-        setCameraModal(false);
-        setPage(page - 1);
-        console.log(page);
-      }
+  const updateData = (data, deleteData = false) => {
+    if (deleteData) {
+      localStorage.removeItem(`id_data_${facilitator?.id}`);
+    } else {
+      localStorage.setItem(`id_data_${facilitator?.id}`, JSON.stringify(data));
     }
-  }, [page, credentials]);
+  };
 
   const uiSchema = {
     dob: {
@@ -167,14 +215,14 @@ export default function App({ ip, id }) {
         hideClearButton: true,
       },
     },
-    /*  DOB : {
+    DOB: {
       "ui:widget": "alt-date",
       "ui:options": {
         yearsRange: yearsRange,
         hideNowButton: true,
         hideClearButton: true,
       },
-    }, */
+    },
 
     qualification: {
       "ui:widget": CustomR,
@@ -218,6 +266,14 @@ export default function App({ ip, id }) {
     ownership: {
       "ui:widget": RadioBtn,
     },
+
+    enrollmentboard: {
+      "ui:widget": RadioBtn,
+    },
+    selectsubjects: {
+      "ui:widget": CheckBoxes,
+    },
+
     // custom radio button with property name
     vo_experience: {
       items: {
@@ -236,7 +292,62 @@ export default function App({ ip, id }) {
       },
     },
   };
+  React.useEffect(async () => {
+    console.log("sandy");
+    let data = await enumRegistryService.getAll();
+    console.log("sandy");
 
+    let dataSchema = updateSchemaEnum(schema, {
+      key: "selectsubjects",
+      arr: data,
+      title: "title",
+      value: "value",
+      // filters: { type: "selectsubjects" },
+    });
+    let newSchema = schema;
+    // console.log(newSchema, dataSchema, "first");
+
+    setSchema(dataSchema);
+    /*   if (schema["properties"]["selectsubjects"]) {
+      newSchema = updateSchemaEnum(newSchema, {
+        key: "selectsubjects",
+        arr: data,
+        title: "title",
+        value: "value",
+        filters: { type: "selectsubjects" },
+      });
+      console.log(newSchema, "second");
+
+      if (newSchema?.properties?.selectsubjects) {
+        let valueIndex = "";
+        newSchema?.properties?.selectsubjects?.enumNames?.forEach(
+          (e, index) => {
+            if (e.match("1")) {
+              valueIndex = newSchema?.properties?.selectsubjects?.enum[index];
+            }
+          }
+        );
+        if (valueIndex !== "" && formData.selectsubjects == valueIndex) {
+          setAlert(t("NO_DATA"));
+        } else {
+          setAlert();
+        }
+      }
+    }
+    if (schema["properties"]["degree"]) {
+      newSchema = updateSchemaEnum(newSchema, {
+        key: "selectsubjects",
+        arr: data,
+        title: "title",
+        value: "value",
+        filters: { type: "selectsubjects" },
+      });
+      console.log(newSchema, "last");
+    } */
+    //setSchema(newSchema);
+  }, [page]);
+  // console.log(schema);
+  console.log(formData);
   const nextPreviewStep = async (pageStape = "n") => {
     setAlert();
     const index = pages.indexOf(page);
@@ -252,8 +363,8 @@ export default function App({ ip, id }) {
         setPage(nextIndex);
         setSchema(properties[nextIndex]);
       } else if (pageStape.toLowerCase() === "n") {
-        await formSubmitUpdate({ ...formData, form_step_number: "6" });
-        setPage("SAVE");
+        await formSubmitUpdate({ ...formData, form_step_number: "13" });
+        setPage("upload");
       } else {
         return true;
       }
@@ -303,17 +414,93 @@ export default function App({ ip, id }) {
     };
   };
 
+  React.useEffect(async () => {
+    if (schema?.properties?.qualification) {
+      const qData = await facilitatorRegistryService.getQualificationAll();
+      let newSchema = schema;
+      if (schema["properties"]["qualification"]) {
+        newSchema = getOptions(newSchema, {
+          key: "qualification",
+          arr: qData,
+          title: "name",
+          value: "id",
+          filters: { type: "qualification" },
+        });
+        if (newSchema?.properties?.qualification) {
+          let valueIndex = "";
+          newSchema?.properties?.qualification?.enumNames?.forEach(
+            (e, index) => {
+              if (e.match("12")) {
+                valueIndex = newSchema?.properties?.qualification?.enum[index];
+              }
+            }
+          );
+          if (valueIndex !== "" && formData.qualification == valueIndex) {
+            setAlert(t("YOU_NOT_ELIGIBLE"));
+          } else {
+            setAlert();
+          }
+        }
+      }
+      if (schema["properties"]["degree"]) {
+        newSchema = getOptions(newSchema, {
+          key: "degree",
+          arr: qData,
+          title: "name",
+          value: "id",
+          filters: { type: "teaching" },
+        });
+      }
+      setSchema(newSchema);
+    }
+
+    if (schema?.properties?.state) {
+      const qData = await geolocationRegistryService.getStates();
+      let newSchema = schema;
+      if (schema["properties"]["state"]) {
+        newSchema = getOptions(newSchema, {
+          key: "state",
+          arr: qData,
+          title: "state_name",
+          value: "state_name",
+        });
+      }
+      newSchema = await setDistric({
+        schemaData: newSchema,
+        state: formData?.state,
+        district: formData?.district,
+        block: formData?.block,
+      });
+      setSchema(newSchema);
+    }
+
+    if (schema?.properties?.device_ownership) {
+      if (formData?.device_ownership == "no") {
+        setAlert(t("YOU_NOT_ELIGIBLE"));
+      } else {
+        setAlert();
+      }
+    }
+  }, [page]);
+
   React.useEffect(() => {
     if (schema1.type === "step") {
       const properties = schema1.properties;
       const newSteps = Object.keys(properties);
+      // console.log(schema1.properties, "sandesh");
+
       setPage(newSteps[0]);
       setSchema(properties[newSteps[0]]);
       setPages(newSteps);
-      let minYear = moment().subtract("years", 30);
+      let minYear = moment().subtract("years", 50);
       let maxYear = moment().subtract("years", 18);
       setYearsRange([minYear.year(), maxYear.year()]);
       setSubmitBtn(t("NEXT"));
+    }
+    if (facilitator?.id) {
+      const data = localStorage.getItem(`id_data_${facilitator?.id}`);
+      const newData = JSON.parse(data);
+      setFormData({ ...newData, ...facilitator });
     }
   }, []);
 
@@ -330,7 +517,6 @@ export default function App({ ip, id }) {
       setSubmitBtn(t("SAVE"));
       setAddBtn(t("ADD_EXPERIENCE"));
     } else {
-      console.log({ ...formData });
       setSubmitBtn(t("SAVE"));
     }
   };
@@ -344,21 +530,28 @@ export default function App({ ip, id }) {
   };
 
   const formSubmitUpdate = async (formData) => {
-    console.log("sent data");
+    const { id } = facilitator;
     if (id) {
-      const data = await enumRegistryService.editProfileById({
+      updateData({}, true);
+      return await facilitatorRegistryService.stepUpdate({
         ...formData,
+        parent_ip: ip?.id,
         id: id,
       });
-      console.log(data, "sent data");
     }
+  };
+  const formSubmitCreate = async (formData) => {
+    console.log(formData);
+    /* await facilitatorRegistryService.create({
+            ...formData,
+            parent_ip: ip?.id,
+        }); */
   };
 
   const goErrorPage = (key) => {
     if (key) {
       pages.forEach((e) => {
-        console.log(e);
-        const data = schema1["properties"]?.[e]["properties"]?.[key];
+        const data = schema1["properties"][e]["properties"][key];
         if (data) {
           setStep(e);
         }
@@ -373,6 +566,16 @@ export default function App({ ip, id }) {
       }
       if (!(data?.mobile > 6666666666 && data?.mobile < 9999999999)) {
         errors.mobile.addError(t("PLEASE_ENTER_VALID_NUMBER"));
+      }
+    }
+    if (data?.aadhar_token) {
+      if (
+        data?.aadhar_token &&
+        !data?.aadhar_token?.match(/^[2-9]{1}[0-9]{3}[0-9]{4}[0-9]{4}$/)
+      ) {
+        errors?.aadhar_token?.addError(
+          `${t("AADHAAR_SHOULD_BE_12_DIGIT_VALID_NUMBER")}`
+        );
       }
     }
     if (data?.dob) {
@@ -515,6 +718,7 @@ export default function App({ ip, id }) {
     setErrors();
     const newData = { ...formData, ...data };
     setFormData(newData);
+    updateData(newData);
     if (id === "root_mobile") {
       if (data?.mobile?.toString()?.length === 10) {
         const result = await userExist({ mobile: data?.mobile });
@@ -593,7 +797,6 @@ export default function App({ ip, id }) {
   };
 
   const onError = (data) => {
-    console.log(data);
     if (data[0]) {
       const key = data[0]?.property?.slice(1);
       goErrorPage(key);
@@ -623,6 +826,7 @@ export default function App({ ip, id }) {
       ["form_step_number"]: parseInt(page) + 1,
     };
     setFormData(newData);
+    updateData(newData);
     if (_.isEmpty(errors)) {
       setStep();
     } else {
@@ -643,6 +847,7 @@ export default function App({ ip, id }) {
       setErrors({ fileSize: t("FILE_SIZE") });
     }
   };
+
   if (cameraUrl) {
     return (
       <Layout
@@ -658,43 +863,27 @@ export default function App({ ip, id }) {
       >
         <VStack py={6} px={4} mb={5} space="6">
           {/* add the profile image */}
-          <H1 color="red.1000">{t("ADD_ID_PHOTOS")}</H1>
+          <H1 color="red.1000">{t("ADD_PROFILE_PHOTO")}</H1>
           <h5 color="red.1000" fontSize="3">
             {t("CLEAR_PROFILE_MESSAGE")}
           </h5>
-
           <Center>
             <Image
               source={{
                 uri: cameraUrl,
-                //show the image
               }}
               alt=""
               size="324px"
             />
           </Center>
-          <h5 color="red.1000" fontSize="2">
-            {t("FRONT_VIEW")}
-          </h5>
           <Button
             variant={"primary"}
             onPress={async (e) => {
-              if (cameraSelection >= 3) {
-                nextPreviewStep();
-                setCameraUrl();
-                setCameraModal(false);
-                pages[pages?.length - 1] === page;
-              } else {
-                setCameraUrl();
-                setCameraModal(true);
-                setCameraSelection(cameraSelection + 1);
-              }
-              /*  await formSubmitUpdate({ ...formData, form_step_number: "5" });
-              console.log(...formData);
-              if (onClick) onClick(cameraUrl); */
+              await formSubmitUpdate({ ...formData, form_step_number: "13" });
+              if (onClick) onClick("success");
             }}
           >
-            {cameraSelection > 3 ? setSubmitBtn(t("SAVE")) : t("SAVE")}
+            {t("SUBMIT")}
           </Button>
           <Button
             variant={"secondary"}
@@ -719,14 +908,76 @@ export default function App({ ip, id }) {
           cameraUrl,
           setCameraUrl: async (url) => {
             setCameraUrl(url);
-            setCameraData([...cameraData, url]);
-            setFormData({ ...formData, ["profile_url"]: cameraData });
+            setFormData({ ...formData, ["profile_url"]: url });
           },
         }}
       />
     );
   }
 
+  if (page === "upload") {
+    return (
+      <Layout
+        _appBar={{ onPressBackButton: (e) => setPage("13"), lang, setLang }}
+        _page={{ _scollView: { bg: "white" } }}
+      >
+        <VStack py={6} px={4} mb={5} space="6">
+          {/* Box removal */}
+          <H1 color="red.1000">{t("JUST_ONE_STEP")}</H1>
+          <H2 color="red.1000">{t("ADD_PROFILE_PHOTO")} -</H2>
+          <Button
+            variant={"primary"}
+            leftIcon={
+              <IconByName
+                name="CameraLineIcon"
+                color="white"
+                size={2}
+                isDisabled
+              />
+            }
+            onPress={(e) => {
+              setCameraUrl();
+              setCameraModal(true);
+            }}
+          >
+            {t("TAKE_PHOTO")}
+          </Button>
+          <VStack space={2}>
+            <input
+              accept="image/*"
+              type="file"
+              style={{ display: "none" }}
+              ref={uplodInputRef}
+              onChange={handleFileInputChange}
+            />
+            <Button
+              leftIcon={<IconByName name="Download2LineIcon" isDisabled />}
+              variant={"secondary"}
+              onPress={(e) => {
+                uplodInputRef?.current?.click();
+              }}
+            >
+              {t("UPLOAD_PHOTO")}
+            </Button>
+            {errors?.fileSize ? (
+              <H2 color="red.400">{errors?.fileSize}</H2>
+            ) : (
+              <React.Fragment />
+            )}
+          </VStack>
+          <Button
+            variant={"primary"}
+            onPress={async (e) => {
+              await formSubmitUpdate({ ...formData, form_step_number: "13" });
+              if (onClick) onClick("success");
+            }}
+          >
+            {t("SKIP_SUBMIT")}
+          </Button>
+        </VStack>
+      </Layout>
+    );
+  }
   const AddButton = ({ icon, iconType, ...btnProps }) => {
     return (
       <Button
