+++ apps/front-end/src/routes/routes.js
@@ -1,8 +1,8 @@
 import Dashboard from "pages/front-end/Dashboard";
 import Home from "pages/front-end/Home";
-import basicDetails from "../pages/front-end/AGDetails/basicDetails";
-import educationDetails from "../pages/front-end/AGDetails/education-details-further-studies/educationDetails";
-import chooseSubjects from "../pages/front-end/AGDetails/choose-subjects/subjectDetails";
+import basicDetails from "../pages/front-end/ag-edit/basicDetails";
+import educationDetails from "../pages/front-end/ag-edit/education-details-further-studies/educationDetails";
+import chooseSubjects from "../pages/front-end/ag-edit/choose-subjects/subjectDetails";
 import Agduplicate from "pages/front-end/ag-form/Agduplicate";
 import Agform from "pages/front-end/ag-form/Agform";
 import Docschecklist from "pages/front-end/ag-form/Docschecklist";
@@ -17,7 +17,7 @@ export default [
     component: Home,
   },
   { path: "/dashboard", component: Dashboard },
-  { path: "/ag/:id", component: basicDetails },
+  { path: "/beneficiary/edit/:id", component: basicDetails },
   { path: "/ag/education/:id", component: educationDetails },
   { path: "/ag/subjects/:id", component: chooseSubjects },
   //add a new route /ag/:ID(param), component:basic details
